Code organization, renaming and merging classes, files, etc
